+++ app/Http/Controllers/Admin/Services/ClientServiceFilesController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Admin\Services;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
